requirements.txt
flask
pymongo[srv]
python-dotenv
joblib
numpy
scikit-learn
werkzeug